requirements.txt
alembic==1.7.5
autopep8==1.6.0
blinker==1.4
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
dnspython==2.2.0
dominate==2.6.0
elasticsearch==7.17.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
pycodestyle==2.8.0
PyJWT==2.3.0
python-dotenv==0.19.2
SQLAlchemy==1.4.31
toml==0.10.2
urllib3==1.26.8
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.1
